sync-github: add the --team-repo flag to easily use a local clone
Before this commit testing GitHub team changes was a pain since you had
to change the contents of the local team repo clone, build its static
API, fire up a local web server and set the TEAM_DATA_BASE_URL
environment variable to point to it.

This commit simplifies that workflow by adding a new --team-repo flag to
sync-github: when the flag is passed along with a path to a team repo
clone, sync-github will automatically build the static API in a
temporary directory and query the data directly from the file system.