promote-release: allow overriding the branch to pull commits from
While this is not needed for the production builds, it might be useful
to publish try builds on dev-static as nightlies, for example to test
how rustup would behave with a manifest change.

This commit adds a PROMOTE_RELEASE_OVERRIDE_BRANCH environment variable
that allows the default branch promote-release pulls from to be
overridden. When the environment variable is not present the current
behavior will not change.