cancelbot: ensure builds from the right repository are fetched
Previously, cancelbot fetched all the builds in the Azure DevOps project
on the requested branch, regardless of the repository they were
originating from. This caused problems once test-on-fork landed, as
multiple repositories were linked to the same Azure DevOps project.

This updates cancelbot to explicitly request builds from the repository
we want, fixing cancelbot when using it with test-on-fork.