Backfill MAJOR.MINOR Rustup channel manifests
Now that https://github.com/rust-lang/rust/pull/76107 has been merged,
new releases will also write their manifests to channel-rust-1.x.toml as
well as channel-rust-1.x.y.toml to enable `rustup install 1.48` to get
the latest patch release in a minor release series.

This commit adds an idempotent script to copy manifests and their
signatures for the last patch release in every minor release series to
the corresponding minor manifest files so that past minor versions will
work with the rustup functionality too.

This script should only need to be run once, but should be safe to run
more than once.

It starts at 1.8 because we don't have manifests for 1.0-1.7, and it ends
with 1.47 because 1.48 will be the first stable release to write out the
1.x channel manifest.